QBitmap: restore nothrow move special members
The user-defined dtor inhibits the move special member
functions.

We cannot do something about it in Qt 5, because the
class is exported (which it shouldn't be), and because
it's polymorphic (which it also shouldn't be), so we
need to supply all the missing member functions manually.

Add a note to remove the inheritance from QPixmap in Qt 6.

Change-Id: I86a1a3fe7b84247b3e604cb008aa17bb673c8468
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>