Add begin-end virtuals to the internal glyph cache
Just having a single fillTexture() is not sufficient for efficient
operation with modern, low-level graphics APIs.

Having a begin-end pair of functions that are invoked before the stream of
calls to fillTexture allow glyph caches to build command lists and defer
command submission until it is known that all pending glyphs have been
processed.

Change-Id: Ifac645f83e9cfb6f246be9c8e79e5aa3bde5758d
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>